vit_floorplan/requirements/base.txt
# -------------------------------
# Core Data Science
# -------------------------------
numpy>=1.26.0
pandas>=2.2.0
scipy>=1.14.0
scikit-learn>=1.5.0

# -------------------------------
# Visualization
# -------------------------------
matplotlib>=3.9.0
seaborn>=0.13.0
plotly>=5.20.0
opencv-python>=4.10.0.84
Pillow>=10.3.0

# -------------------------------
# Deep Learning (CUDA 11.8)
# -------------------------------
torch==2.5.1+cu118
torchvision==0.20.1+cu118
torchaudio==2.5.1+cu118
pytorch-lightning>=2.3.0
--extra-index-url https://download.pytorch.org/whl/cu118

# -------------------------------
# Transformers & Models
# -------------------------------
transformers>=4.43.0
timm>=1.0.3
segmentation-models-pytorch>=0.3.3

# -------------------------------
# Experiment Tracking
# -------------------------------
mlflow>=2.12.0
wandb>=0.17.0
optuna>=3.5.0

# -------------------------------
# Data Management & Config
# -------------------------------
hydra-core>=1.3.2
omegaconf>=2.3.0
dvc>=3.40.0
pyyaml>=6.0.1

# -------------------------------
# Utilities
# -------------------------------
tqdm>=4.66.3
python-dotenv>=1.0.1
pydantic>=2.8.0
loguru>=0.7.2

# -------------------------------
# Development & Testing
# -------------------------------
pytest>=8.2.0
pytest-cov>=5.0.0
black>=24.3.0
ruff>=0.5.6

# -------------------------------
# Hugging Face Hub
# -------------------------------
huggingface-hub>=0.25.0
sentencepiece>=0.2.0
tokenizers>=0.19.0
